+++ NEWS.md
@@ -1,4 +1,4 @@
-# broom.helpers (development version)
+# broom.helpers 1.19.0
 
 **Deprecated function**